perf: /mypage Suspense 分割 + Sanity 全クエリのキャッシュ化（BON-316, BON-318）
- /mypage を Suspense × 3 に分割し、各セクションを独立 Server Component 化
  - ヘッダー+タブが TTFB 直後に表示（旧 ~814ms → ~10ms）
  - bookmarks/history が重い lessons クエリを待たず先行表示
  - 画像最適化（unoptimized 削除、img → Image 置換）
- src/lib/sanity.ts の 30 Sanity GROQ 関数を unstable_cache でラップ
  - タグは document type 名（lesson, article, quest, feedback, blogPost, training, guide, event, roadmap）
  - revalidate: 3600 + タグ無効化のハイブリッド
  - 計測実証: cache hit で 1ms（getAllLessonsWithArticleIds 764ms → 1ms = 約 760倍）
- scripts/measure-mypage-sanity.ts: Sanity GROQ 単体実時間の計測スクリプト
