+++ src/app/mypage/page.tsx
@@ -1,88 +1,68 @@
+import { Suspense } from "react";
 import { Metadata } from "next";
 import { redirect } from "next/navigation";
 import { getCurrentUser } from "@/lib/subscription";
-import { getAllLessonsWithArticleIds } from "@/lib/sanity";
-import { getBookmarkedArticles } from "@/lib/services/bookmarks";
-import { getViewHistory } from "@/lib/services/viewHistory";
-import {
-  getMultipleLessonProgress,
-  getMultipleLessonStatus,
-} from "@/lib/services/progress";
-import MyPageClient from "./MyPageClient";
+import { MyPageShell, type TabId } from "./MyPageShell";
+import { ProgressSection } from "./sections/ProgressSection";
+import { BookmarksSection } from "./sections/BookmarksSection";
+import { HistorySection } from "./sections/HistorySection";
+import { ProgressSkeleton } from "./skeletons/ProgressSkeleton";
+import { BookmarksSkeleton } from "./skeletons/BookmarksSkeleton";
+import { HistorySkeleton } from "./skeletons/HistorySkeleton";
 
 export const metadata: Metadata = {
   title: "マイページ",
   description: "あなたの学習ダッシュボード",
 };
 
-export default async function MyPage() {
-  const user = await getCurrentUser();
+const VALID_TABS = new Set<TabId>(["all", "progress", "favorite", "history"]);
 
+function parseTab(value: string | string[] | undefined): TabId {
+  const t = typeof value === "string" ? value : "all";
+  return VALID_TABS.has(t as TabId) ? (t as TabId) : "all";
+}
+
+export default async function MyPage({
+  searchParams,
+}: {
+  searchParams: Promise<{ tab?: string | string[] }>;
+}) {
+  console.time("[mypage] shell ready (until first byte)");
+  const user = await getCurrentUser();
   if (!user) {
     redirect("/login?redirectTo=/mypage");
   }
 
-  // データを並列取得
-  const [lessons, bookmarkedArticles, viewHistory] = await Promise.all([
-    getAllLessonsWithArticleIds(),
-    getBookmarkedArticles(user.id),
-    getViewHistory(user.id),
-  ]);
+  const params = await searchParams;
+  const tab = parseTab(params.tab);
 
-  // レッスンの進捗を一括取得
-  const lessonProgressInput = lessons.map((lesson) => ({
-    lessonId: lesson._id,
-    articleIds: lesson.articleIds || [],
-  }));
+  const showProgress = tab === "all" || tab === "progress";
+  const showBookmarks = tab === "all" || tab === "favorite";
+  const showHistory = tab === "all" || tab === "history";
 
-  const [progressMap, statusMap] = await Promise.all([
-    getMultipleLessonProgress(lessonProgressInput, user.id),
-    getMultipleLessonStatus(
-      lessons.map((l) => l._id),
-      user.id
-    ),
-  ]);
+  const progressMode: "preview" | "full" = tab === "progress" ? "full" : "preview";
+  const bookmarksMode: "preview" | "full" = tab === "favorite" ? "full" : "preview";
+  const historyMode: "preview" | "full" = tab === "history" ? "full" : "preview";
 
-  // レッスン + 進捗データを統合
-  const lessonsWithProgress = lessons.map((lesson) => {
-    const progress = progressMap[lesson._id] || null;
-    const lessonStatus = statusMap[lesson._id] || "not_started";
-
-    // 最初の未完了記事を特定
-    let firstIncompleteArticle = null;
-    if (lesson.quests) {
-      const completedIds = progress?.completedArticleIds || [];
-      for (const quest of lesson.quests) {
-        if (!quest.articles) continue;
-        for (const article of quest.articles) {
-          if (!completedIds.includes(article._id)) {
-            firstIncompleteArticle = {
-              title: article.title,
-              slug: article.slug?.current || "",
-            };
-            break;
-          }
-        }
-        if (firstIncompleteArticle) break;
-      }
-    }
-
-    return {
-      _id: lesson._id,
-      title: lesson.title,
-      slug: lesson.slug,
-      iconImageUrl: lesson.iconImageUrl,
-      progress,
-      lessonStatus,
-      firstIncompleteArticle,
-    };
-  });
+  console.timeEnd("[mypage] shell ready (until first byte)");
 
   return (
-    <MyPageClient
-      bookmarkedArticles={bookmarkedArticles}
-      viewHistory={viewHistory}
-      lessonsWithProgress={lessonsWithProgress}
-    />
+    <MyPageShell activeTab={tab}>
+      {showProgress && (
+        <Suspense fallback={<ProgressSkeleton mode={progressMode} />}>
+          <ProgressSection userId={user.id} mode={progressMode} />
+        </Suspense>
+      )}
+      {showBookmarks && (
+        <Suspense fallback={<BookmarksSkeleton mode={bookmarksMode} />}>
+          <BookmarksSection userId={user.id} mode={bookmarksMode} />
+        </Suspense>
+      )}
+      {showHistory && (
+        <Suspense fallback={<HistorySkeleton mode={historyMode} />}>
+          <HistorySection userId={user.id} mode={historyMode} />
+        </Suspense>
+      )}
+    </MyPageShell>
   );
 }
